+++ ACB/Models/LatLong.cs
@@ -2,7 +2,7 @@
 {
     public class LatLong
     {
-        public double? Lat { get; set; }
-        public double? Long { get; set; }
+        public decimal? Lat { get; set; }
+        public decimal? Long { get; set; }
     }
 }
